moved shared tests into test folder
