+++ packages/shared/tests/utils/get-fields-from-template.test.ts
@@ -1,4 +1,4 @@
-import { getFieldsFromTemplate } from './get-fields-from-template';
+import { getFieldsFromTemplate } from '../../src/utils/get-fields-from-template';
 
 describe('getFieldsFromTemplate', () => {
 	it('returns an empty array when passed null', () => {
