+++ packages/shared/tests/utils/move-in-array.test.ts
@@ -1,4 +1,4 @@
-import { moveInArray } from './move-in-array';
+import { moveInArray } from '../../src/utils/move-in-array';
 
 describe('moveInArray', () => {
 	const testArray = [1, 2, 3, 4, 5, 6];
